+++ src/main/java/net/silentchaos512/gear/command/RandomGearCommand.java
@@ -16,7 +16,6 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.text.TranslationTextComponent;
 import net.minecraftforge.registries.ForgeRegistries;
-import net.silentchaos512.gear.SilentGear;
 import net.silentchaos512.gear.api.item.ICoreItem;
 import net.silentchaos512.gear.init.ModItems;
 import net.silentchaos512.gear.util.GearGenerator;
@@ -26,7 +25,7 @@ import java.util.Collection;
 
 public final class RandomGearCommand {
     private static final SuggestionProvider<CommandSource> itemIdSuggestions = (context, builder) ->
-            ISuggestionProvider.func_212476_a(ModItems.gearClasses.keySet().stream().map(SilentGear::getId), builder);
+            ISuggestionProvider.func_212476_a(ModItems.gearClasses.keySet().stream(), builder);
 
     private RandomGearCommand() {}